Color class is created
Color class is required for Draught (see #2).
It might be that rework will be needed for Color and for Coord.
It is a plan to use union of objects of these classes in Draught.
Most probably it will break Color and Coord functionality.
However, rework to use bit masks will help