[Engine] Draught class was created.
It is representation of piece of checkers.
It contains coordinates on a board, color and king flag in
just 8 bits.
Now it can make itself as copy of another Draught or
from parts (Coord, Color, king flag).
It is able to make move of draught to any squere (no check that
it is board still though).
It is able to make draught king and get if king or not.
It can compare two draughts:
1. Check if they are the same or not (==, !=)
2. Say which one is farther from 'a1' square (compare coords only).
   X has bigger priority here, so 'b5' is less than 'c1' but bigger
   than 'a7'.

I think at this stage I did everything in class Draught what planned.
If it will require rework, it will be done for another (probably new)
ticket.
This commit closes #2.