[Engine] Board does not check kind of draught put into king row
Currently the Board allows to put Man on King row but it is wrong.
The (failed) test putManOnKingRow was created to remind about this.

The test put was modified to be sure that after fix it will be still
possible to put King on King row.

However, this bug is not very important stuff, so I will fix it then.
See #16 meta ticket for suddenly found bugs.