[Engine] Modification of basic class for Actions (see #5)
After some thinking I decided that it is better to have some factory
which knows how to create specific action by Board and Move.

However, I'm to lazy to implement the factory right now.
But I can probably good approach: give an action a responsibility
to create (clone) itself if it is right action.

The changes:
1. Pure virtual factory method added to ActionAtBoard
2. transformIntoActions return ActionAtBoard::Ptr now
3. Board + Move reworked to throw exception ErrorUnknown if
   transformIntoActions return null pointer.

No action list in Board + Move means that:
1. Simple Jump action most probably will be implemented as Aggregator
   of Move action and Take Away action;
2. Complex Jump action most probably will be implemented as Aggregator
   of Simple Jumps.