[Engine] 'Previous' implementation for Simple Move (#see 5)
It is action based implementation for Simple Move.
All SimpleMove tests are passed.
Set of algorithms which created in RulesOfGame and
AmericanCheckersActionFactory will be reused (and/or reworked)
during planned refactoring.

Also some of functions added into RulesOfGame are tested indirectly.
I've added TODO to test them, but do not want to waste time on it right now.
I'm not really sure that this code will not be moved / reworked a lot
during planned refactoring.
So this task goes to suddenly found bugs (see #16)