[Engine] Rework in Loop::Walker to allow childs to use self pointer
It was wound that  shared pointer to self is needed in player.
And it shall be shared pointer to Player class instead of
shared pointer to Walker class to avoid casting.

This experimental rework is an attempt to choose type of self pointer
to store in Walker.

Since it is experemental, it is on dev branch.

If Player implementation will be good enough, it will be merged to
master. Or Player will be implemented in different manner.