[Engine] JustOwnMoveAheadMind is implemented (closes #8)
One more stupid mind is implemented.
In fact, it works the same as RandomChoiceMind for most of cases.
The only exception if Jumps with different scores exist.
In this case it chooses one of the best score jumps.
Sevedral plays with RandomChoiceMind shows that they have almost the
same chance to win or loose.

Tests are added.