[Logger] Exception logging is added (see #24)
log_*_throw can be used to log the moment when exception it thrown.
log_catch_throw is used to log an exception and re-throw it.
log_catch_ignore is used to ignore caught exception but log it just in case.

All macro log additional information for std::exception based classes.